backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.25.2
django-allauth==0.57.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
